Allow any existing appender to be run in its own thread.
The Async facade puts a queue between the appender and the main message
queue, so that slow appenders do not hold up logging.
The main logging thread now also uses this shared facade.
Eventually this could allow specific logger instances to write to
different processors so that the log messages don’t have to be global
all the time.
Also opens the door to not use a separate thread for logging if
required.