Always re-raise with LoadError
Fixes #138
Fixes sorbet/sorbet#1820

It's a best practice to re-raise a LoadError exception when catching a
LoadError exception, so that scripts can attempt to be resilient in the
face of a LoadError. Having to swallow all RuntimeExceptions is the
alternative, and it often makes things more brittle, rather than less.