Fix inconsistencies in timezone handling for elasticsearch appender
Elasticsearch appender is using a raw formatter by default, and the
formatter is using Time#utc method, which modifies the time object
in-place. The result is resolving different index name in non-UTC
timezones around midnight.

This commit fixes it by forcing the formatter to operate on a duplicate
of the time object, so it's not mutated. As a result, one test had to be fixed.