Fix `flush` on `AsyncBatch` appender
When `flush` is called on a `AsyncBatch` appender `batch` is called
after `process_message` has finished (which handles the `flush` call).
This means that when `flush` returns the log messages are not actually
written yet and if the process exits at this point the logs are lost.

This was observed to lose log messages in an AWS Lambda environment.