Adds support for newer versions of Elasticsearch that removed _type in Elasticsearch Appender.
I tested this code with 6.x, 7.x and 8.x. For testing with the
different versions, I think we would need to add different gemfiles
and run the tests with each version, but I don't know if that's worth
it or something you'd like to have since it'd be quite an overhead.

I added a helper to the tests. For versions < 8.0.0,
Elasticsearch::Transport::Client and Elasticsearch::Client could be
used to call `bulk`, but since we extracted Transport into its own
library, you need to call Elasticsearch::Client for bulk.

Addresses #225